requirements.txt
pydantic
langchain
langchain-core
langchain-community
langchain-groq
langgraph
python-dotenv
setuptools